Fix unsafe API
emmalloc changes memory on free() so need to call it only when finished
working with the data.

Technically we can return {frameRef, data} instead of just data and
stroing lastFrameRef inside a class, but allocating several uncompressed
frames might quickly consume all available WebAssembly memory so it's
not a good idea anyway.